energy_analytics/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
schedule>=1.2.0
aiohttp>=3.8.0

# Data processing
requests>=2.31.0
urllib3>=2.0.0
pytz>=2023.3

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Monitoring and logging
python-json-logger>=2.0.0

# Data validation
pydantic>=2.0.0
